test(gateway): split strict probe coverage
