+++ src/gateway/server-http.probe.test.ts
@@ -903,34 +903,6 @@ describe("gateway probe endpoints", () => {
     });
   });
 
-  it("makes strict liveness draining-aware without inheriting channel readiness", async () => {
-    let failing = ["gateway-draining"];
-    const getReadiness: ReadinessChecker = () => ({
-      ready: false,
-      failing,
-      uptimeMs: 999,
-    });
-
-    await withGatewayServer({
-      prefix: "probe-healthz-strict-draining",
-      resolvedAuth: AUTH_NONE,
-      overrides: { getReadiness },
-      run: async (server) => {
-        const draining = await sendGatewayRequest(server, { path: "/healthz?strict=1" });
-        expect(draining.res.statusCode).toBe(503);
-        expect(JSON.parse(draining.getBody())).toEqual({
-          live: false,
-          phase: "shutting_down",
-        });
-
-        failing = ["discord"];
-        const channelFailure = await sendGatewayRequest(server, { path: "/healthz?strict=1" });
-        expect(channelFailure.res.statusCode).toBe(200);
-        expect(JSON.parse(channelFailure.getBody())).toEqual({ ok: true, status: "live" });
-      },
-    });
-  });
-
   it("serves liveness probes before loading gateway config or resolving auth", async () => {
     const getRuntimeConfig = vi.fn(() => {
       throw new Error("config load blocked");
